camera: Pull chessboard calibration camera bits into seperate module
The chessboard module is seperate from the rest of the calibration bits.
It is good to seperate it as a module because it has a small set of
dependencies and is purely functional.

This will make it easier to refactor, clean-up and reuse.
